基于 Redis 封装分布式数据操作 分布式锁和分布式计数器

+++ 11redis/redis-lock/src/main/java/com/catcher/redislock/count/RedisCount.java
@@ -3,7 +3,6 @@ package com.catcher.redislock.count;
 import org.springframework.stereotype.Service;
 import redis.clients.jedis.Jedis;
 import redis.clients.jedis.JedisPool;
-import redis.clients.jedis.params.SetParams;
 
 import java.util.UUID;
 
@@ -14,13 +13,13 @@ public class RedisCount {
     private final JedisPool jedisPool = new JedisPool("127.0.0.1", 6379);
 
 
-    public long setIncrBy(String key, long increment, long times){
+    public long setIncrBy(String key, long increment, int times){
         long result = 0;
         Jedis connnect = jedisPool.getResource();
         long end = System.currentTimeMillis() + times;
         result = connnect.incrBy(key, increment);
         if (result <= 1 && times != 0){
-            connnect.expire(key, times);
+            connnect.expire(key, (int) times);
         }
         jedisPool.returnResource(connnect);
         return result;
